Potential fix for code scanning alert no. 48: Multiplication result converted to larger type

+++ src/LiveBackgroundRemovalLite/MainFilter/RenderingContext.cpp
@@ -250,8 +250,10 @@ void RenderingContext::videoRender()
 					bb.y * segmenterRoi_.height / selfieSegmenter_->getHeight() + segmenterRoi_.y);
 				sourceRoi_.width = static_cast<std::uint32_t>(bb.width * segmenterRoi_.width /
 									      selfieSegmenter_->getWidth());
-				sourceRoi_.height = static_cast<std::uint32_t>(bb.height * segmenterRoi_.height /
-									       selfieSegmenter_->getHeight());
+				sourceRoi_.height = static_cast<std::uint32_t>(
+					static_cast<std::uint64_t>(bb.height) *
+					static_cast<std::uint64_t>(segmenterRoi_.height) /
+					static_cast<std::uint64_t>(selfieSegmenter_->getHeight()));
 				sourceRoi_.y = static_cast<std::uint32_t>(
 					static_cast<std::uint64_t>(bb.y) *
 						static_cast<std::uint64_t>(segmenterRoi_.height) /
